Refactor: move JsonUploader logic from AppContext to JsonUploader.tsx
