+++ components/AppContext.tsx
@@ -1,7 +1,7 @@
 import { createContext, useContext, useState, useRef, useEffect } from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import styles from '../styles/Workspace.module.css';
-import { getAdditionalData, editProject, postProjectBtn, postProjectVideo, postProjectAnnotation, postVideoAnnotation, getProjectAnnotation } from '@/utils/requests';
+import { getAdditionalData, postVideoAnnotation, getProjectAnnotation } from '@/utils/requests';
 import { clearUnfinishedAnnotation } from '@/utils/utils';
 import { UploaderType } from '@/types/misc';
 import BtnGroup from './BtnGroup';
@@ -51,16 +51,22 @@ saveAnnotation: boolean,
 skeletonLandmark: string,
 undo: number,
 updateAnnotationChart: boolean,
-uploader: UploadFile,
+uploader: UploaderType,
 useEraser: boolean,
 videoAdditionalFieldsConfig: {},
 videoData: {},
 videoId: string,
-annotationRef: {},
+annotationRef: annoRefType,
 lastFrameNumForIntervalAnnoRef: number,
 lastFrameNumForIntervalErasingRef: number,
 realFpsRef: number,
 videoMetaRef: {},
+}
+
+type annoRefType = {
+  current: {
+
+  }
 }
 
 const StatesContext = createContext<StatesType | undefined>(undefined);
@@ -306,137 +312,7 @@ export default function StatesProvider({children}: AppContextProps) {
             setAdditionalData(additionalDataForChart);
         }
     }        
-
-    // There's another useEffect for uploader inside AnnotationChart.  Should that be combined here?  Or not?  
-    useEffect(() => {
-        if (uploader?.type && uploader?.file?.originFileObj) { //checking in case undefined
-            saveAnnotationAndUpdateStates(true); 
-            const reader = new FileReader();
-            reader.onload = (e) => onReaderLoad(e, uploader.type);
-            reader.readAsText(uploader.file.originFileObj);
-        }
-    }, [uploader])
-
-    function onReaderLoad(e, type: string){
-        
-        const obj = JSON.parse(e.target.result);
-        if (type === 'annotation') {
-            /**
-             * {
-             *      projectId: str,
-             *      videos: [str, str, ],
-             *      annotations: [{}, {}, ...]
-             * }
-             */
-            
-            if (!projectId) {
-                setModalInfo('Please upload the project configuration data first');
-                setModalInfoOpen(true);
-            } else if (obj.projectId !== projectId) {
-                setModalInfo('The project id in the uploaded data does not match the current project id. Please upload the project configuration file first.')
-                setModalInfoOpen(true);
-            } else {
-                Modal.confirm({
-                    content: 'Upload and save/update the uploaded annotation data to database?\nThis will overwrite the data in database.',
-                    onOk: () => {confirmSaveUploadedAnnotationToDB(obj)},
-                });
-            }
-            
-        } else {
-            /**
-             * {
-             *      projectId: str,
-                    projectName: str,
-                    (description: str,) optional
-                    btnConfigData: {},
-                    videos: {}
-                }
-             */
-            if (projectId) {
-                Modal.confirm({
-                    title: 'Alert',
-                    content: 'The current project data will be replaced!\nThe uploaded configuration data will be saved to database. This may overwrite the existing data in database.',
-                    onOk: ()=>{confirmUploadConfiguration(obj)},
-                });
-            } else {
-                Modal.confirm({
-                    title: 'Alert',
-                    content: 'The uploaded configuration data will be saved to database. This may overwrite the existing data in database.',
-                    onOk: ()=>{confirmUploadConfiguration(obj)},
-                });
-            }
-        }
-    }
-
-    async function confirmUploadConfiguration(obj) {
-        const projectObj = {
-            projectId: obj.projectId, 
-            projectName: obj.projectName,
-            description: obj.description
-        }
-        const projectRes = await editProject(projectObj);
-        if (projectRes['error']) {
-            setGlobalInfo('Editting project configuration data in DB failed.');
-            return
-        } 
-        setProjectId(obj.projectId);
-        setProjectData({projectId: obj.projectId, projectName: obj.projectName, description: obj.description});
-
-        
-        const btns = Object.keys(obj.btnConfigData).map((id)=>{
-            const btnGroupObj = {...obj.btnConfigData[id]};
-            btnGroupObj.btnGroupId = id;
-            return btnGroupObj;
-        })
-        const btnDataForDB = {
-            btnGroups: btns,
-            projectId: obj.projectId
-        }
-        let btnRes = await postProjectBtn(btnDataForDB);
-        if (btnRes['error']) {
-            setGlobalInfo('Saving btn configuration data to DB failed.');
-            return
-        } 
-
-        const videos = Object.keys(obj.videos).map((id)=>{
-            const videoObj = {...obj.videos[id]};
-            videoObj.videoId = id;
-            return videoObj;
-        })
-        const videoDataForDB = {
-            videos: videos,
-            projectId: obj.projectId
-        }
-        let videoRes = await postProjectVideo(videoDataForDB);
-        if (videoRes['error']) {
-            setGlobalInfo('Saving video data to DB failed.');
-            return
-        }
-        
-        setBtnConfigData(obj.btnConfigData ? {...obj.btnConfigData} : {});
-        setVideoData(obj.videos ? {...obj.videos} : {});
-    }
-
-    async function confirmSaveUploadedAnnotationToDB(data) {
-        const res = await postProjectAnnotation({...data});
-        if (res['error']) {
-            setGlobalInfo('Saving annotation data to DB failed.');
-        } else {
-            if ((data.videos.filter(v => v===videoId).length>0) && Number.isInteger(frameNum)) {
-                const videoAnnotations = data.annotations.filter(anno => anno.videoId === videoId);
-                const forAnnoRef = {};
-                videoAnnotations.forEach(anno => {
-                    if (!forAnnoRef[anno.frameNum]) {
-                        forAnnoRef[anno.frameNum] = {};
-                    }
-                    forAnnoRef[anno.frameNum][anno.id] = anno;
-                })
-                annotationRef.current = forAnnoRef;
-                getFrameAnnotationFromRefAndSetState();
-            } 
-            setResetAnnotationChart(true);
-        }
-    }
+ 
 
     useEffect(()=> {
       if (downloadConfig) {
@@ -589,12 +465,12 @@ export default function StatesProvider({children}: AppContextProps) {
             }
 
             if (Object.keys(newFrameAnno).length > 0) {
-                const firstAnno = Object.values(newFrameAnno)[0];
-                if (savePrevFrame && firstAnno.frameNum === frameNum-1) {
-                    annotationRef.current[frameNum-1] = newFrameAnno; 
-                } else if (!savePrevFrame && firstAnno.frameNum === frameNum) {
-                    annotationRef.current[frameNum] = newFrameAnno; 
-                }
+              const firstAnno = Object.values(newFrameAnno)[0];
+              if (savePrevFrame && frameNum !== undefined && firstAnno.frameNum === frameNum-1) {
+                  annotationRef.current[frameNum-1] = newFrameAnno; 
+              } else if (!savePrevFrame && firstAnno.frameNum === frameNum && frameNum !== undefined) {
+                  annotationRef.current[frameNum] = newFrameAnno; 
+              }
             } 
           
     }
@@ -681,8 +557,10 @@ export default function StatesProvider({children}: AppContextProps) {
     }
 
     function getFrameAnnotationFromRefAndSetState() {
+      if (frameNum !== undefined){
         const frameAnno = annotationRef.current[frameNum]??{};
         setFrameAnnotation({...frameAnno});
+      }
     }
   
     return (
